refactor(app): extract the table-browsing and sidebar wiring
Opening a table into a browse query, the pagination footer, the Mongo
filter bar and JSON tree, refresh, expanding schema headers and table
columns, the sidebar tree filter and the Items/Queries/History switch
move to `wire::browse`.

The `wire::*` calls now sit together at the end of `main` instead of
where each block used to be. They only register callbacks, so the order
between them carries no meaning — the requirement is just that the state
and shared closures exist first.
